Move fill_pool
